feat(recall): organize generated insights and quality
Keep durable context and generated reports together under Recall, with explicit report scope that survives tab changes and deep links. Move deterministic session signals to Quality so model-written output is no longer mixed with measured facts.
