+++ docs/screenshots/update-generated-assets-branch.sh
@@ -38,8 +38,7 @@ expected_assets=(
   "screenshots/import-modal-chatgpt.png"
   "screenshots/import-modal-claude.png"
   "screenshots/in-session-search.png"
-  "screenshots/insight-content.png"
-  "screenshots/insights.png"
+  "screenshots/quality.png"
   "screenshots/layout-compact.png"
   "screenshots/layout-stream.png"
   "screenshots/machine-labels.png"
@@ -48,6 +47,7 @@ expected_assets=(
   "screenshots/project-breakdown.png"
   "screenshots/publish-modal.png"
   "screenshots/recall-corpus.png"
+  "screenshots/recall-generated-insights.png"
   "screenshots/recent-edits.png"
   "screenshots/resync-modal.png"
   "screenshots/search-grouped.png"
