+++ docs/scripts/check_built_site.py
@@ -22,7 +22,8 @@ ROUTES = [
     "/mcp/",
     "/token-usage/",
     "/chat-import/",
-    "/insights/",
+    "/quality/",
+    "/recall/",
     "/commands/",
     "/stats/",
     "/session-api/",
