+++ docs/zensical.toml
@@ -23,7 +23,7 @@ nav = [
   {"MCP Server" = "mcp.md"},
   {"Token Usage & Costs" = "token-usage.md"},
   {"Chat Import" = "chat-import.md"},
-  {"Session Insights" = "insights.md"},
+  {"Quality" = "quality.md"},
   {"CLI Reference" = "commands.md"},
   {"Session Export" = "session-export.md"},
   {"Reporting Export" = "reporting-export.md"},
