+++ frontend/messages/en.json
@@ -21,7 +21,7 @@
   "nav_trends": "Trends",
   "nav_recall": "Recall",
   "nav_pinned": "Pinned",
-  "nav_insights": "Insights",
+  "nav_quality": "Quality",
   "nav_trash": "Trash",
   "nav_search_sessions": "Search sessions...",
   "nav_search_sessions_shortcut": "Search sessions ({shortcut})",
@@ -906,7 +906,7 @@
   "settings_search_keywords_remote_access": "remote server connection auth token network",
   "settings_loading": "Loading settings...",
   "settings_date_ranges_title": "Date ranges",
-  "settings_date_ranges_description": "Share date selections across Sessions, Usage, Activity, Trends, and Insights.",
+  "settings_date_ranges_description": "Share date selections across Sessions, Usage, Activity, Trends, and Quality.",
   "settings_date_ranges_link": "Link date ranges across pages",
   "settings_resync_title_unavailable": "Full resync unavailable in read-only mode",
   "settings_resync_unavailable_hint": "Unavailable while connected to a read-only backend",
@@ -1025,7 +1025,7 @@
   "activity_insight_generate_insight": "Generate insight",
   "activity_insight_generation_failed": "Generation failed",
   "activity_insight_title": "Activity Insight",
-  "activity_insight_open_insights_page": "Open in Insights page",
+  "activity_insight_open_generated": "Open in Generated insights",
   "activity_insight_agent_cli_title": "Agent CLI used to generate the insight",
   "activity_insight_insight_agent": "Insight agent",
   "activity_insight_loading": "Loading insight…",
@@ -1282,6 +1282,19 @@
   "session_recall_evidence_range": "Messages {start}–{end}",
   "session_recall_jump_evidence": "Jump to transcript messages {start}–{end}",
   "recall_page_title": "Recall",
+  "recall_page_workspace_subtitle": "Browse durable knowledge and generated reports.",
+  "recall_page_tabs_label": "Recall workspace",
+  "recall_page_tab_corpus": "Corpus",
+  "recall_page_tab_generated": "Generated insights",
+  "recall_generated_subtitle": "Generate reports from an explicit session scope or browse saved reports.",
+  "recall_generated_date_range": "Date range",
+  "recall_generated_project": "Project",
+  "recall_generated_session_agent": "Session agent",
+  "recall_generated_select_session_agent": "Filter report sessions by agent",
+  "recall_generated_archive_label": "Saved generated insights",
+  "recall_generated_export": "Export",
+  "recall_generated_publish": "Publish",
+  "recall_generated_publish_secret": "Secret",
   "recall_page_subtitle": "Browse distilled entries and extraction coverage.",
   "recall_page_search_placeholder": "Search Recall entries…",
   "recall_page_clear_search": "Clear Recall search",
@@ -1299,7 +1312,6 @@
   "recall_page_trigger_label": "Trigger",
   "recall_page_uncertainty_label": "Uncertainty",
   "recall_page_loading": "Loading Recall entries…",
-  "recall_page_unavailable_read_only": "Recall is unavailable in read-only mode.",
   "recall_page_empty": "No Recall entries match these filters.",
   "recall_page_error": "Could not load Recall entries.",
   "recall_page_load_more": "Load more",
@@ -1491,7 +1503,6 @@
   "insights_page_filter_scopes": "Filter scopes...",
   "insights_page_filter_by_scope": "Filter insights by session scope",
   "insights_page_no_matching_scopes": "No matching scopes",
-  "insights_page_refresh": "Refresh insights",
   "insights_page_rule_based": "Rule-based",
   "insights_page_next_actions": "Next actions",
   "insights_page_deterministic_recommendations": "Deterministic Recommendations",
@@ -1500,8 +1511,9 @@
   "insights_page_scored_facts": "Scored facts",
   "insights_page_quality_patterns": "Quality Patterns",
   "insights_page_deterministic_counts": "Deterministic counts from persisted session signals for {range}.",
-  "insights_page_insights_help_intro": "Deterministic sections are computed from session data, while generated insights are separate context text.",
-  "insights_page_insights_help_docs": "Read Insights docs",
+  "quality_page_refresh": "Refresh quality",
+  "quality_page_help_intro": "Quality signals and recommendations are computed directly from session data.",
+  "quality_page_help_docs": "Read Quality docs",
   "insights_page_could_not_load": "Could not load deterministic insights.",
   "insights_page_retry": "Retry",
   "insights_page_no_scored_data": "No scored quality data for this range.",
@@ -1620,8 +1632,6 @@
   "insights_page_failures": "{count} failures",
   "insights_page_generated": "Generated",
   "insights_page_separate_from_facts": "Separate from scored facts",
-  "insights_page_generated_archive": "Generated Insights Archive",
-  "insights_page_generated_archive_hint": "Saved generated text is shown separately and does not affect deterministic quality scores. Creation remains out of scope for this deterministic phase.",
   "insights_page_template_label": "Template",
   "insights_page_filter_templates": "Filter templates...",
   "insights_page_select_template": "Select template",
