+++ frontend/messages/fr.json
@@ -21,7 +21,7 @@
   "nav_trends": "Tendances",
   "nav_recall": "Rappel",
   "nav_pinned": "Épinglés",
-  "nav_insights": "Analyses",
+  "nav_quality": "Qualité",
   "nav_trash": "Corbeille",
   "nav_search_sessions": "Rechercher des sessions...",
   "nav_search_sessions_shortcut": "Rechercher des sessions ({shortcut})",
@@ -906,7 +906,7 @@
   "settings_search_keywords_remote_access": "distant serveur connexion authentification jeton réseau remote server connection auth token network",
   "settings_loading": "Chargement des paramètres...",
   "settings_date_ranges_title": "Plages de dates",
-  "settings_date_ranges_description": "Partager la sélection de dates entre Sessions, Consommation, Activité, Tendances et Analyses.",
+  "settings_date_ranges_description": "Partager la sélection de dates entre Sessions, Consommation, Activité, Tendances et Qualité.",
   "settings_date_ranges_link": "Lier les plages de dates entre les pages",
   "settings_resync_title_unavailable": "Resynchronisation complète indisponible en mode lecture seule",
   "settings_resync_unavailable_hint": "Indisponible lorsqu'on est connecté à un backend en lecture seule",
@@ -1025,7 +1025,7 @@
   "activity_insight_generate_insight": "Générer une analyse",
   "activity_insight_generation_failed": "Échec de la génération",
   "activity_insight_title": "Analyse d'activité",
-  "activity_insight_open_insights_page": "Ouvrir dans la page Analyses",
+  "activity_insight_open_generated": "Ouvrir dans les analyses générées",
   "activity_insight_agent_cli_title": "CLI d'agent utilisée pour générer l'analyse",
   "activity_insight_insight_agent": "Agent d'analyse",
   "activity_insight_loading": "Chargement de l'analyse…",
@@ -1282,6 +1282,19 @@
   "session_recall_evidence_range": "Messages {start}–{end}",
   "session_recall_jump_evidence": "Aller aux messages {start}–{end} de la transcription",
   "recall_page_title": "Rappel",
+  "recall_page_workspace_subtitle": "Parcourez les connaissances durables et les rapports générés.",
+  "recall_page_tabs_label": "Espace Rappel",
+  "recall_page_tab_corpus": "Corpus",
+  "recall_page_tab_generated": "Analyses générées",
+  "recall_generated_subtitle": "Générez des rapports à partir d’un périmètre de sessions explicite ou parcourez les rapports enregistrés.",
+  "recall_generated_date_range": "Plage de dates",
+  "recall_generated_project": "Projet",
+  "recall_generated_session_agent": "Agent de session",
+  "recall_generated_select_session_agent": "Filtrer les sessions du rapport par agent",
+  "recall_generated_archive_label": "Analyses générées enregistrées",
+  "recall_generated_export": "Exporter",
+  "recall_generated_publish": "Publier",
+  "recall_generated_publish_secret": "Secret",
   "recall_page_subtitle": "Parcourir les entrées distillées et la couverture d’extraction.",
   "recall_page_search_placeholder": "Rechercher dans le rappel…",
   "recall_page_clear_search": "Effacer la recherche de rappel",
@@ -1299,7 +1312,6 @@
   "recall_page_trigger_label": "Déclencheur",
   "recall_page_uncertainty_label": "Incertitude",
   "recall_page_loading": "Chargement des entrées de rappel…",
-  "recall_page_unavailable_read_only": "Le rappel n’est pas disponible en mode lecture seule.",
   "recall_page_empty": "Aucune entrée de rappel ne correspond à ces filtres.",
   "recall_page_error": "Impossible de charger les entrées de rappel.",
   "recall_page_load_more": "Afficher plus",
@@ -1490,7 +1502,6 @@
   "insights_page_filter_scopes": "Filtrer les portées...",
   "insights_page_filter_by_scope": "Filtrer les analyses par portée de session",
   "insights_page_no_matching_scopes": "Aucune portée correspondante",
-  "insights_page_refresh": "Actualiser les analyses",
   "insights_page_rule_based": "Basé sur des règles",
   "insights_page_next_actions": "Actions à mener",
   "insights_page_deterministic_recommendations": "Recommandations déterministes",
@@ -1499,8 +1510,9 @@
   "insights_page_scored_facts": "Faits notés",
   "insights_page_quality_patterns": "Motifs de qualité",
   "insights_page_deterministic_counts": "Décomptes déterministes issus des signaux de session persistés pour {range}.",
-  "insights_page_insights_help_intro": "Les sections déterministes sont calculées à partir des données de session, tandis que les analyses générées sont un texte de contexte distinct.",
-  "insights_page_insights_help_docs": "Lire la documentation des analyses",
+  "quality_page_refresh": "Actualiser la qualité",
+  "quality_page_help_intro": "Les signaux de qualité et les recommandations sont calculés directement à partir des données de session.",
+  "quality_page_help_docs": "Lire la documentation sur la qualité",
   "insights_page_could_not_load": "Impossible de charger les analyses déterministes.",
   "insights_page_retry": "Réessayer",
   "insights_page_no_scored_data": "Aucune donnée de qualité notée pour cette plage.",
@@ -1619,8 +1631,6 @@
   "insights_page_failures": "{count} échecs",
   "insights_page_generated": "Générée",
   "insights_page_separate_from_facts": "Distinct des faits notés",
-  "insights_page_generated_archive": "Archive des analyses générées",
-  "insights_page_generated_archive_hint": "Le texte généré enregistré est affiché séparément et n'influence pas les scores de qualité déterministes. Sa création reste hors du périmètre de cette phase déterministe.",
   "insights_page_template_label": "Modèle",
   "insights_page_filter_templates": "Filtrer les modèles...",
   "insights_page_select_template": "Sélectionner un modèle",
