+++ frontend/messages/ko.json
@@ -21,7 +21,7 @@
   "nav_trends": "트렌드",
   "nav_recall": "리콜",
   "nav_pinned": "고정됨",
-  "nav_insights": "인사이트",
+  "nav_quality": "품질",
   "nav_trash": "휴지통",
   "nav_search_sessions": "세션 검색...",
   "nav_search_sessions_shortcut": "세션 검색 ({shortcut})",
@@ -881,7 +881,7 @@
   "settings_search_keywords_remote_access": "원격 서버 연결 인증 토큰 네트워크 remote server connection auth token network",
   "settings_loading": "설정을 불러오는 중...",
   "settings_date_ranges_title": "날짜 범위",
-  "settings_date_ranges_description": "세션, 사용량, 활동, 트렌드, 인사이트 페이지에서 선택한 날짜 범위를 공유합니다.",
+  "settings_date_ranges_description": "세션, 사용량, 활동, 트렌드, 품질 페이지에서 선택한 날짜 범위를 공유합니다.",
   "settings_date_ranges_link": "페이지 간 날짜 범위 연결",
   "settings_resync_title_unavailable": "읽기 전용 모드에서는 전체 재동기화를 사용할 수 없습니다",
   "settings_resync_unavailable_hint": "읽기 전용 백엔드에 연결된 동안에는 사용할 수 없습니다",
@@ -1000,7 +1000,7 @@
   "activity_insight_generate_insight": "인사이트 생성",
   "activity_insight_generation_failed": "생성 실패",
   "activity_insight_title": "활동 인사이트",
-  "activity_insight_open_insights_page": "인사이트 페이지에서 열기",
+  "activity_insight_open_generated": "생성된 인사이트에서 열기",
   "activity_insight_agent_cli_title": "인사이트 생성에 사용된 에이전트 CLI",
   "activity_insight_insight_agent": "인사이트 에이전트",
   "activity_insight_loading": "인사이트를 불러오는 중…",
@@ -1252,6 +1252,19 @@
   "session_recall_evidence_range": "메시지 {start}–{end}",
   "session_recall_jump_evidence": "대화 기록 메시지 {start}–{end}(으)로 이동",
   "recall_page_title": "리콜",
+  "recall_page_workspace_subtitle": "지속 가능한 지식과 생성된 보고서를 살펴봅니다.",
+  "recall_page_tabs_label": "리콜 작업 공간",
+  "recall_page_tab_corpus": "코퍼스",
+  "recall_page_tab_generated": "생성된 인사이트",
+  "recall_generated_subtitle": "명시적인 세션 범위에서 보고서를 생성하거나 저장된 보고서를 살펴봅니다.",
+  "recall_generated_date_range": "날짜 범위",
+  "recall_generated_project": "프로젝트",
+  "recall_generated_session_agent": "세션 에이전트",
+  "recall_generated_select_session_agent": "에이전트별 보고서 세션 필터링",
+  "recall_generated_archive_label": "저장된 생성 인사이트",
+  "recall_generated_export": "내보내기",
+  "recall_generated_publish": "게시",
+  "recall_generated_publish_secret": "비공개",
   "recall_page_subtitle": "추출된 항목과 추출 범위를 살펴봅니다.",
   "recall_page_search_placeholder": "리콜 항목 검색…",
   "recall_page_clear_search": "리콜 검색 지우기",
@@ -1269,7 +1282,6 @@
   "recall_page_trigger_label": "트리거",
   "recall_page_uncertainty_label": "불확실성",
   "recall_page_loading": "리콜 항목을 불러오는 중…",
-  "recall_page_unavailable_read_only": "읽기 전용 모드에서는 리콜을 사용할 수 없습니다.",
   "recall_page_empty": "이 필터와 일치하는 리콜 항목이 없습니다.",
   "recall_page_error": "리콜 항목을 불러올 수 없습니다.",
   "recall_page_load_more": "더 보기",
@@ -1459,7 +1471,6 @@
   "insights_page_filter_scopes": "범위 필터링...",
   "insights_page_filter_by_scope": "세션 범위별로 인사이트 필터링",
   "insights_page_no_matching_scopes": "일치하는 범위 없음",
-  "insights_page_refresh": "인사이트 새로고침",
   "insights_page_rule_based": "규칙 기반",
   "insights_page_next_actions": "다음 작업",
   "insights_page_deterministic_recommendations": "결정론적 권장 사항",
@@ -1468,8 +1479,9 @@
   "insights_page_scored_facts": "채점 기반",
   "insights_page_quality_patterns": "품질 패턴",
   "insights_page_deterministic_counts": "{range} 기간의 저장된 세션 신호에서 산출한 결정론적 집계입니다.",
-  "insights_page_insights_help_intro": "결정론적 섹션은 세션 데이터에서 계산되며, 생성된 인사이트는 별도의 컨텍스트 텍스트입니다.",
-  "insights_page_insights_help_docs": "인사이트 문서 읽기",
+  "quality_page_refresh": "품질 새로고침",
+  "quality_page_help_intro": "품질 신호와 권장 사항은 세션 데이터에서 직접 계산됩니다.",
+  "quality_page_help_docs": "품질 문서 읽기",
   "insights_page_could_not_load": "결정론적 인사이트를 불러올 수 없습니다.",
   "insights_page_retry": "재시도",
   "insights_page_no_scored_data": "이 기간에는 채점된 품질 데이터가 없습니다.",
@@ -1585,8 +1597,6 @@
   "insights_page_failures": "실패 {count}건",
   "insights_page_generated": "생성됨",
   "insights_page_separate_from_facts": "채점된 사실과는 별개",
-  "insights_page_generated_archive": "생성된 인사이트 아카이브",
-  "insights_page_generated_archive_hint": "저장된 생성 텍스트는 별도로 표시되며 결정론적 품질 점수에 영향을 주지 않습니다. 생성 기능은 이 결정론적 단계의 범위에서 제외됩니다.",
   "insights_page_template_label": "템플릿",
   "insights_page_filter_templates": "템플릿 필터링...",
   "insights_page_select_template": "템플릿 선택",
