+++ frontend/messages/zh-CN.json
@@ -21,7 +21,7 @@
   "nav_trends": "趋势",
   "nav_recall": "召回",
   "nav_pinned": "已固定",
-  "nav_insights": "洞察",
+  "nav_quality": "质量",
   "nav_trash": "回收站",
   "nav_search_sessions": "搜索会话...",
   "nav_search_sessions_shortcut": "搜索会话 ({shortcut})",
@@ -879,7 +879,7 @@
   "settings_search_keywords_remote_access": "远程 服务器 连接 鉴权 令牌 网络 remote server connection auth token network",
   "settings_loading": "正在加载设置...",
   "settings_date_ranges_title": "日期范围",
-  "settings_date_ranges_description": "在会话、用量、活动、趋势和洞察页面之间共享日期选择。",
+  "settings_date_ranges_description": "在会话、用量、活动、趋势和质量页面之间共享日期选择。",
   "settings_date_ranges_link": "跨页面关联日期范围",
   "settings_resync_title_unavailable": "只读模式下无法完整重新同步",
   "settings_resync_unavailable_hint": "连接到只读后端时不可用",
@@ -998,7 +998,7 @@
   "activity_insight_generate_insight": "生成洞察",
   "activity_insight_generation_failed": "生成失败",
   "activity_insight_title": "活动洞察",
-  "activity_insight_open_insights_page": "在洞察页面中打开",
+  "activity_insight_open_generated": "在生成的洞察中打开",
   "activity_insight_agent_cli_title": "用于生成洞察的代理 CLI",
   "activity_insight_insight_agent": "洞察代理",
   "activity_insight_loading": "正在加载洞察…",
@@ -1250,6 +1250,19 @@
   "session_recall_evidence_range": "消息 {start}–{end}",
   "session_recall_jump_evidence": "跳转到转录消息 {start}–{end}",
   "recall_page_title": "召回",
+  "recall_page_workspace_subtitle": "浏览持久知识和生成的报告。",
+  "recall_page_tabs_label": "召回工作区",
+  "recall_page_tab_corpus": "语料库",
+  "recall_page_tab_generated": "生成的洞察",
+  "recall_generated_subtitle": "从明确的会话范围生成报告，或浏览已保存的报告。",
+  "recall_generated_date_range": "日期范围",
+  "recall_generated_project": "项目",
+  "recall_generated_session_agent": "会话代理",
+  "recall_generated_select_session_agent": "按代理筛选报告会话",
+  "recall_generated_archive_label": "已保存的生成洞察",
+  "recall_generated_export": "导出",
+  "recall_generated_publish": "发布",
+  "recall_generated_publish_secret": "私密",
   "recall_page_subtitle": "浏览提炼条目和提取覆盖情况。",
   "recall_page_search_placeholder": "搜索召回条目…",
   "recall_page_clear_search": "清除召回搜索",
@@ -1267,7 +1280,6 @@
   "recall_page_trigger_label": "触发条件",
   "recall_page_uncertainty_label": "不确定性",
   "recall_page_loading": "正在加载召回条目…",
-  "recall_page_unavailable_read_only": "只读模式下无法使用召回。",
   "recall_page_empty": "没有符合这些筛选条件的召回条目。",
   "recall_page_error": "无法加载召回条目。",
   "recall_page_load_more": "加载更多",
@@ -1457,7 +1469,6 @@
   "insights_page_filter_scopes": "筛选范围...",
   "insights_page_filter_by_scope": "按会话范围筛选洞察",
   "insights_page_no_matching_scopes": "无匹配范围",
-  "insights_page_refresh": "刷新洞察",
   "insights_page_rule_based": "基于规则",
   "insights_page_next_actions": "下一步行动",
   "insights_page_deterministic_recommendations": "确定性建议",
@@ -1466,8 +1477,9 @@
   "insights_page_scored_facts": "评分事实",
   "insights_page_quality_patterns": "质量模式",
   "insights_page_deterministic_counts": "基于 {range} 持久化会话信号的确定性统计。",
-  "insights_page_insights_help_intro": "确定性部分由会话数据计算得出，生成洞察是分开展示的说明文本。",
-  "insights_page_insights_help_docs": "查看 Insights 文档",
+  "quality_page_refresh": "刷新质量",
+  "quality_page_help_intro": "质量信号和建议直接根据会话数据计算。",
+  "quality_page_help_docs": "查看质量文档",
   "insights_page_could_not_load": "无法加载确定性洞察。",
   "insights_page_retry": "重试",
   "insights_page_no_scored_data": "此范围内无评分质量数据。",
@@ -1583,8 +1595,6 @@
   "insights_page_failures": "{count} 次失败",
   "insights_page_generated": "已生成",
   "insights_page_separate_from_facts": "与评分事实分离",
-  "insights_page_generated_archive": "生成洞察存档",
-  "insights_page_generated_archive_hint": "已保存的生成文本单独显示，不影响确定性质量评分。在当前确定性阶段，生成功能仍不在范围内。",
   "insights_page_template_label": "模板",
   "insights_page_filter_templates": "筛选模板...",
   "insights_page_select_template": "选择模板",
