+++ frontend/messages/zh-TW.json
@@ -21,7 +21,7 @@
   "nav_trends": "趨勢",
   "nav_recall": "召回",
   "nav_pinned": "已釘選",
-  "nav_insights": "洞察",
+  "nav_quality": "品質",
   "nav_trash": "垃圾桶",
   "nav_search_sessions": "搜尋對話...",
   "nav_search_sessions_shortcut": "搜尋對話 ({shortcut})",
@@ -879,7 +879,7 @@
   "settings_search_keywords_remote_access": "遠端 伺服器 連線 驗證 權杖 網路 remote server connection auth token network",
   "settings_loading": "正在載入設定...",
   "settings_date_ranges_title": "日期範圍",
-  "settings_date_ranges_description": "在對話、用量、活動、趨勢和洞察頁面之間共享日期選擇。",
+  "settings_date_ranges_description": "在對話、用量、活動、趨勢和品質頁面之間共享日期選擇。",
   "settings_date_ranges_link": "跨頁面連結日期範圍",
   "settings_resync_title_unavailable": "唯讀模式下無法完整重新同步",
   "settings_resync_unavailable_hint": "連線到唯讀後端時不可用",
@@ -998,7 +998,7 @@
   "activity_insight_generate_insight": "生成洞察",
   "activity_insight_generation_failed": "生成失敗",
   "activity_insight_title": "活動洞察",
-  "activity_insight_open_insights_page": "在洞察頁面中開啟",
+  "activity_insight_open_generated": "在生成的洞察中開啟",
   "activity_insight_agent_cli_title": "用於生成洞察的代理 CLI",
   "activity_insight_insight_agent": "洞察代理",
   "activity_insight_loading": "正在載入洞察…",
@@ -1250,6 +1250,19 @@
   "session_recall_evidence_range": "訊息 {start}-{end}",
   "session_recall_jump_evidence": "跳至逐字稿訊息 {start}-{end}",
   "recall_page_title": "召回",
+  "recall_page_workspace_subtitle": "瀏覽持久知識與生成的報告。",
+  "recall_page_tabs_label": "召回工作區",
+  "recall_page_tab_corpus": "語料庫",
+  "recall_page_tab_generated": "生成的洞察",
+  "recall_generated_subtitle": "從明確的對話範圍生成報告，或瀏覽已儲存的報告。",
+  "recall_generated_date_range": "日期範圍",
+  "recall_generated_project": "專案",
+  "recall_generated_session_agent": "對話代理",
+  "recall_generated_select_session_agent": "依代理篩選報告對話",
+  "recall_generated_archive_label": "已儲存的生成洞察",
+  "recall_generated_export": "匯出",
+  "recall_generated_publish": "發布",
+  "recall_generated_publish_secret": "私密",
   "recall_page_subtitle": "瀏覽提煉項目和擷取涵蓋情況。",
   "recall_page_search_placeholder": "搜尋召回項目…",
   "recall_page_clear_search": "清除召回搜尋",
@@ -1267,7 +1280,6 @@
   "recall_page_trigger_label": "觸發條件",
   "recall_page_uncertainty_label": "不確定性",
   "recall_page_loading": "正在載入召回項目…",
-  "recall_page_unavailable_read_only": "唯讀模式下無法使用召回。",
   "recall_page_empty": "沒有符合這些篩選條件的召回項目。",
   "recall_page_error": "無法載入召回項目。",
   "recall_page_load_more": "載入更多",
@@ -1457,7 +1469,6 @@
   "insights_page_filter_scopes": "篩選範圍...",
   "insights_page_filter_by_scope": "按對話範圍篩選洞察",
   "insights_page_no_matching_scopes": "無匹配範圍",
-  "insights_page_refresh": "重整洞察",
   "insights_page_rule_based": "基於規則",
   "insights_page_next_actions": "下一步行動",
   "insights_page_deterministic_recommendations": "確定性建議",
@@ -1466,8 +1477,9 @@
   "insights_page_scored_facts": "評分事實",
   "insights_page_quality_patterns": "品質模式",
   "insights_page_deterministic_counts": "基於 {range} 持久化對話信號的確定性統計。",
-  "insights_page_insights_help_intro": "確定性部分由對話資料計算得出，生成洞察是分開展示的說明文本。",
-  "insights_page_insights_help_docs": "查看 Insights 文件",
+  "quality_page_refresh": "重新整理品質",
+  "quality_page_help_intro": "品質訊號和建議直接根據工作階段資料計算。",
+  "quality_page_help_docs": "查看品質文件",
   "insights_page_could_not_load": "無法載入確定性洞察。",
   "insights_page_retry": "重試",
   "insights_page_no_scored_data": "此範圍內無評分品質資料。",
@@ -1583,8 +1595,6 @@
   "insights_page_failures": "{count} 次失敗",
   "insights_page_generated": "已生成",
   "insights_page_separate_from_facts": "與評分事實分離",
-  "insights_page_generated_archive": "生成洞察存檔",
-  "insights_page_generated_archive_hint": "已儲存的生成文本單獨顯示，不影響確定性品質評分。在目前確定性階段，生成功能仍不在範圍內。",
   "insights_page_template_label": "模板",
   "insights_page_filter_templates": "篩選模板...",
   "insights_page_select_template": "選擇模板",
