+++ frontend/src/App.svelte
@@ -39,7 +39,6 @@
 
 <script lang="ts">
   import { onMount, untrack } from "svelte";
-  import { EmptyState } from "@kenn-io/kit-ui";
   import AppHeader from "./lib/components/layout/AppHeader.svelte";
   import ThreeColumnLayout from "./lib/components/layout/ThreeColumnLayout.svelte";
   import SessionBreadcrumb from "./lib/components/layout/SessionBreadcrumb.svelte";
@@ -62,7 +61,7 @@
   import ActivityPage from "./lib/components/activity/ActivityPage.svelte";
   import TrendsPage from "./lib/components/trends/TrendsPage.svelte";
   import RecallPage from "./lib/components/recall/RecallPage.svelte";
-  import InsightsPage from "./lib/components/insights/InsightsPage.svelte";
+  import QualityPage from "./lib/components/quality/QualityPage.svelte";
   import PinnedPage from "./lib/components/pinned/PinnedPage.svelte";
   import TrashPage from "./lib/components/trash/TrashPage.svelte";
   import RecentEditsPage from "./lib/components/recentedits/RecentEditsPage.svelte";
@@ -100,14 +99,6 @@
   } from "./lib/stores/sessionRouteParams.js";
 
   let globalAuthToken: string = $state("");
-  const recallModeKnown = $derived(
-    sync.serverVersion !== null ||
-      (settings.loaded && settings.error === null),
-  );
-  const recallUnavailable = $derived(
-    sync.readOnly || settings.readOnly,
-  );
-
   function handleGlobalAuth() {
     const token = globalAuthToken.trim();
     if (!token) return;
@@ -678,19 +669,11 @@
   </div>
 {:else if router.route === "recall"}
   <div class="page-scroll">
-    {#if recallUnavailable}
-      <div class="recall-route-state">
-        <EmptyState title={m.recall_page_unavailable_read_only()} />
-      </div>
-    {:else if recallModeKnown}
-      <RecallPage />
-    {:else}
-      <p class="recall-route-state">{m.recall_page_loading()}</p>
-    {/if}
+    <RecallPage />
   </div>
-{:else if router.route === "insights"}
+{:else if router.route === "quality"}
   <div class="page-scroll">
-    <InsightsPage />
+    <QualityPage />
   </div>
 {:else if router.route === "pinned"}
   <div class="page-scroll">
@@ -811,12 +794,6 @@
   .settings-page-host {
     display: flex;
     overflow: hidden;
-  }
-
-  .recall-route-state {
-    padding: 32px 24px;
-    color: var(--text-muted);
-    font-size: 13px;
   }
 
   /* kit-ui-check-ignore: undo toast carries an inline restore action; kit-ui FlashBanner only supports text+dismiss today, so replacing this would change the delete/undo workflow. */
