+++ frontend/src/lib/components/quality/qualityPatterns.test.ts
