+++ frontend/src/lib/components/quality/qualityPatterns.ts
