+++ frontend/src/lib/stores/analytics.svelte.ts
@@ -147,7 +147,7 @@ class AnalyticsStore {
   private abortControllers: Partial<Record<Panel, AbortController>> = {};
   // Scope key of the cached `signals`: the Analytics-only filters (model plus
   // the heatmap drill-down) the cached data was fetched with. Used to drop the
-  // cache when a fetch crosses the Analytics / Insights boundary, where those
+  // cache when a fetch crosses the Analytics / Quality boundary, where those
   // filters do not exist.
   private signalsScope: string | null = null;
 
@@ -286,7 +286,7 @@ class AnalyticsStore {
   setAutomatedScope(scope: AutomatedScope) {
     this.automatedScope = scope;
     this.includeAutomated = scope !== "human";
-    this.fetchSignalsForInsights();
+    this.fetchSignalsForQuality();
   }
 
   clearRecentlyActive() {
@@ -463,8 +463,8 @@ class AnalyticsStore {
   }
 
   signalEvidenceParams(): AnalyticsParams {
-    // Insights-only drilldown: omit the Analytics model filter so signal
-    // evidence matches the unscoped Insights signal facts (the Insights page
+    // Quality-only drilldown: omit the Analytics model filter so signal
+    // evidence matches the unscoped Quality signal facts (the Quality page
     // has no model control).
     return this.filterParams({ includeModel: false });
   }
@@ -750,9 +750,9 @@ class AnalyticsStore {
     opts: { includeModel?: boolean } = {},
   ): Promise<FetchResult> {
     const includeModel = opts.includeModel ?? true;
-    // `signals` is a cache shared by the Analytics page and the Insights page.
-    // Key it by the filters that exist on Analytics but not Insights: the model
-    // and the heatmap drill-down (date/day/hour), which fetchSignalsForInsights
+    // `signals` is a cache shared by the Analytics page and the Quality page.
+    // Key it by the filters that exist on Analytics but not Quality: the model
+    // and the heatmap drill-down (date/day/hour), which fetchSignalsForQuality
     // clears. When this fetch's scope differs from the cached one, drop the
     // cache so another scope's signals are never shown while the fetch is in
     // flight or retained if it fails; a matching scope keeps the in-place
@@ -780,14 +780,14 @@ class AnalyticsStore {
     );
   }
 
-  async fetchSignalsForInsights() {
+  async fetchSignalsForQuality() {
     this.rollDates();
     this.selectedDate = null;
     this.selectedDow = null;
     this.selectedHour = null;
-    // The Insights page has no model control and the model filter is an
+    // The Quality page has no model control and the model filter is an
     // Analytics-only scope; omit it so a model selected on Analytics does not
-    // silently narrow the Insights signal facts.
+    // silently narrow the Quality signal facts.
     await this.fetchSignals({ includeModel: false });
   }
