+++ frontend/src/lib/stores/analyticsPageDates.ts
@@ -1,7 +1,7 @@
 import { rollingRange } from "../utils/dates.js";
 import type { PanelDateState } from "./yokedDates.svelte.js";
 
-export type AnalyticsDatePage = "sessions" | "insights";
+export type AnalyticsDatePage = "sessions" | "quality";
 
 export interface RetainedAnalyticsPageDate {
   state: PanelDateState;
@@ -10,7 +10,7 @@ export interface RetainedAnalyticsPageDate {
 
 const DEFAULT_WINDOW_DAYS = 365;
 
-// Sessions and Insights share analytics results and filters, but an opted-out
+// Sessions and Quality share analytics results and filters, but an opted-out
 // date selection and its explicit-intent provenance must survive navigation
 // without becoming the other page's selection. Rolling ranges are
 // rematerialized so retained presets stay fresh.
